+++ src/statistics/mod.rs
@@ -1,2 +1,3 @@
 pub mod stat;
 pub mod rand;
+pub mod dist;